Filter link types
Link types selectable in the drop down are now restricted to types that
are editable by the user. This will make it impossible to create link
states that will fail when saving the new work item. There are still two
types shown when using the current filtering that aren't shown in the
link tap (Reported against builds & attachments). I couldn't figure out
why these don't show up in the link tab, but it seems like they are
allowed values.